backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
celery
redis
kombu
google-generativeai
pandas
scikit-learn
python-dotenv
pytest
httpx
